test(pdf): split workspace-relative and native-provider suites out of the main file
Keeps pdf-tool.test.ts under the oxlint max-lines budget (894 lines) by
moving the workspace-relative resolver cases and the native-provider
execution-path block into focused suites; coverage is unchanged.
